Move settings UI shell into package
