+++ Packages/CmuxSettingsUI/Sources/CmuxSettingsUI/SettingsCardNote.swift
@@ -1,0 +1,18 @@
+import SwiftUI
+
+public struct SettingsCardNote: View {
+    public let text: String
+
+    public init(_ text: String) {
+        self.text = text
+    }
+
+    public var body: some View {
+        Text(text)
+            .font(.caption)
+            .foregroundColor(.secondary)
+            .padding(.horizontal, 14)
+            .padding(.vertical, 8)
+            .frame(maxWidth: .infinity, alignment: .leading)
+    }
+}
