+++ Sources/SettingsCardNote.swift
@@ -1,18 +1,3 @@
-import SwiftUI
+import CmuxSettingsUI
 
-struct SettingsCardNote: View {
-    let text: String
-
-    init(_ text: String) {
-        self.text = text
-    }
-
-    var body: some View {
-        Text(text)
-            .font(.caption)
-            .foregroundColor(.secondary)
-            .padding(.horizontal, 14)
-            .padding(.vertical, 8)
-            .frame(maxWidth: .infinity, alignment: .leading)
-    }
-}
+typealias SettingsCardNote = CmuxSettingsUI.SettingsCardNote
